feat\!: unify RithmicResponse.error as Option<RithmicError> for 2.0
Replace the display-only `error: Option<String>` field with a typed
`Option<RithmicError>`, eliminating the `Some("")` sentinel and giving
callers a single ergonomic check that covers both transport failures and
protocol rejections.

Breaking changes:
- `RithmicResponse.error` is now `Option<RithmicError>`
- Removed `RithmicResponse::is_error()`, `is_connection_issue()`,
  `request_error()`, `request_rejection()` — use `.error.is_some()` and
  `err.is_connection_issue()` instead
- `RpCodeClassification::error_message()` and `get_error()` removed;
  replaced by `into_error()` / `classify_rp_code_error()`
- `emit_connection_health_event` / `fail_connection_and_drain` now take
  a single `RithmicError` parameter; message is derived via the new
  `RithmicError::as_connection_message()` helper

New on `RithmicError`: `HeartbeatTimeout`, `ForcedLogout(String)`,
`is_connection_issue()`, `as_connection_message()`.
